+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.0.0 - Unreleased
+2.0.0 - 2021-05-16
 ------------------
 
 - Add ``timeout`` option for LDAP connections. Defaults to 10 seconds.